refactor(dead): extract shared oracle-gate and set-query helpers
The cargo and staticcheck oracle files each carried their own copy of
the Sense-side subset-gate plumbing, and queries.go repeated the same
scan-into-set loop five times.

- New gate.go holds deadSymbols/falseDeads plus the shared
  symRef/DeadSymbol join types; GoDead/RustDead become aliases and the
  exported functions thin wrappers, keeping each oracle file to what
  actually differs (how the oracle runs and parses).
- querySet[T] replaces the five hand-rolled single-column set queries.

Behavior-preserving; existing tests unchanged and green.
